+++ HashStruct.h
@@ -9,7 +9,7 @@
 using namespace std;
 typedef struct lista
 {
-   int chave;
+	long long int chave;
    int prox;
   bool posVazia;
 } lista;
@@ -20,14 +20,14 @@ private:
   lista *tabela;
   int m;
   long long int comparacoes;  
-  int enlacamento(int chave);
-  int divisao(int chave);
-  void encadCol(int chave);
+  long long int enlacamento(long long int chave);
+  long long int divisao(long long int chave);
+  void encadCol(long long int chave);
 
 
 public:
   HashStruct(int tam);
-  void inserir(int chave);
+  void inserir(long long int chave);
   void exibirTabela();
   void salvarArquivo(string nome);
   ~HashStruct();
